iteration 2 with XGboost model
