+++ src/models/treebased/model.py
@@ -6,8 +6,7 @@ import joblib
 import numpy as np
 from numpy.typing import NDArray
 from scipy.stats import halfnorm
-from sklearn.base import BaseEstimator
-from xgboost import XGBClassifier as XGB
+from xgboost import XGBClassifier
 
 from src.page_classes import (
     PageClasses,
@@ -65,82 +64,22 @@ class TreeBasedModel:
         self.model = joblib.load(Path(model_path))
 
 
-class XGBClassifier(BaseEstimator):
-    """Base XGBoost classifier wrapped in base estimator."""
-
-    def __init__(self, objective: str, num_class: int, **kwargs):
-        """Initialisation of the XGBoost classifier.
-
-        Args:
-            objective (str): Objective function for XGBoost.
-            num_class (int): Numebr of classes (with OOD).
-            kwargs (dict): Additional parameters.
-        """
-        self.id2label = id2label
-        self.model = XGB(objective=objective, num_class=num_class, **kwargs)
-
-    def fit(self, X: NDArray[np.float64], y: NDArray[np.float64]) -> Self:
-        """Fit model using SKLearn BaseEstimator.
-
-        Args:
-            X (NDArray[np.float64]): Training features.
-            y (NDArray[np.float64]): Training label.
-
-        Returns:
-            Self: Fitted model
-        """
-        self.is_fitted_ = True
-        self.model.fit(X, y)
-        return self
-
-    def predict(self, X: NDArray[np.float64]) -> NDArray[np.int64]:
-        """Predict classes based on input features.
-
-        Args:
-            X (NDArray[np.float64]): Features to predict.
-
-        Returns:
-            NDArray[np.int64]: Predicted classes
-        """
-        return self.model.predict(X)
-
-    def get_tree_model(self) -> XGB:
-        """Get inner tree model.
-
-        Returns:
-            XGB: Returned model
-        """
-        return self.model
-
-    def get_model_feature_importances(self) -> XGB:
-        """Get inner tree model.
-
-        Returns:
-            XGB: Returned model
-        """
-        if not hasattr(self.model, "feature_importances_"):
-            raise ValueError("Model does not have feature importances. Ensure it is a tree-based model.")
-        return self.model.feature_importances_
-
-
 class XGBOODClassifier(XGBClassifier):
-    """Out of distribution (OOD) XGBoost classifier wrapped in base estimator."""
+    """Out of distribution (OOD) XGBoost classifier."""
 
-    def __init__(self, objective: str, num_class: int, prob_ood: float = 0.95, **kwargs):
+    def __init__(self, prob_ood: float = 0.95, **kwargs):
         """Initialisation of the XGBoostOOD classifier.
 
         Args:
-            objective (str): Objective function for XGBoost.
-            num_class (int): Numebr of classes (with OOD).
+            num_class (int): Number of classes (with OOD).
             prob_ood (float, optional): Probability for OOD detection. Defaults to 0.95.
             kwargs (dict): Additional parameters.
         """
-        super().__init__(objective, num_class, **kwargs)
-
         self.id_ood = label2id[PageClasses.UNKNOWN]
         self.prob_ood = prob_ood
-        self.num_class = num_class - 1
-        self.thresholds = np.zeros_like(self.num_class, dtype=np.float64)
+        self.thresholds = np.zeros(kwargs.get("num_class") - 1, dtype=np.float64)
+        # XGBoost's init can overwrite attrs set after it
+        super().__init__(**kwargs)
 
     def _estimate_thresholds(self, X: NDArray[np.float64], y: NDArray[np.float64]) -> NDArray[np.float64]:
         """Estimate thresholds for each classed in OOD detection.
@@ -152,43 +91,44 @@ class XGBOODClassifier(XGBClassifier):
         Returns:
             NDArray[np.float64]: _description_
         """
-        thresholds = np.zeros(self.num_class, dtype=np.float64)
-        for c in range(self.num_class):
+        thresholds = np.zeros_like(self.thresholds)
+        for c in range(len(thresholds)):
             # Fit probability for given class to half normal distribution
-            Xc_proba = self.model.predict_proba(X[y == c, :])
+            Xc_proba = super().predict_proba(X[y == c, :])
             _, sigma = halfnorm.fit(1 - Xc_proba[:, c], floc=0)
             # Define threshold as prob_ood confidence
             thresholds[c] = 1 - halfnorm.ppf(self.prob_ood, scale=sigma)
         return thresholds
 
-    def fit(self, X: NDArray[np.float64], y: NDArray[np.float64]) -> Self:
-        """Fit model using SKLearn BaseEstimator.
+    def fit(self, X: NDArray[np.float64], y: NDArray[np.float64], **kwargs) -> Self:
+        """Fit model.
 
         Args:
             X (NDArray[np.float64]): Training features.
             y (NDArray[np.float64]): Training label.
+            kwargs (dict): Additional parameters.
 
         Returns:
             Self: Fitted model
         """
-        self.is_fitted_ = True
         # Step 1: Fit XGBoost with all classes except OOD
-        self.model.fit(X[self.id_ood != y, :], y[self.id_ood != y])
+        super().fit(X[self.id_ood != y, :], y[self.id_ood != y], **kwargs)
         # Step 2: Estimate OOD threhsiold based on class distribution
         self.thresholds = self._estimate_thresholds(X, y)
         return self
 
-    def predict(self, X: NDArray[np.float64]) -> NDArray[np.int64]:
+    def predict(self, X: NDArray[np.float64], **kwargs) -> NDArray[np.int64]:
         """Predict classes based on input features.
 
         Args:
             X (NDArray[np.float64]): Features to predict.
+            kwargs (dict): Additional parameters.
 
         Returns:
             NDArray[np.int64]: Predicted classes
         """
         # Compute probability over all classes and get argmax/max
-        y_proba = self.model.predict_proba(X)
+        y_proba = super().predict_proba(X, **kwargs)
         y_label = y_proba.argmax(axis=1)
         y_label_th = y_proba.max(axis=1)
         # Replace prediction where threshold is not meet
